Azure/requirements.txt
blinker==1.9.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
Flask==3.1.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
fonttools==4.55.2
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.3
numpy==2.1.3
packaging==24.2
paho-mqtt==2.1.0
pillow==11.0.0
pycryptodome==3.21.0
pymssql==2.3.2
pyodbc==5.2.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.17.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
Werkzeug==3.1.3
